requirements.txt
imageio==2.37.0
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
tifffile==2025.5.24
tzdata==2025.2
